xi-nub: work-in-progress implementation of xi nub server
a nub is the term used for cached process activation, whereby a
background process is automatically started if not already running,
letting us cache indices that are expensive to create every invocation.

this initial implementation contains the client and server protocol
to perform searches using a persistent process that hosts the
unicode character database index. the caching of the rabin-karp
index reduces the search time by a factor of 30x.

this is a progress commit as the nub code does not yet support
robust error handling and automatic invocation of the nub process.
the eventual implementation will start the nub server automatically.

the nub interface is intended to allow a command-line program to be
split into a typical short lived command invocation and a background
process that is created on demand, which is communicated with for a
stateful function, in this case, cached unicode database indices.

the nub IO interface, is similar to libuv, however, the interface
uses continuations instead of polling because this more closely
models IO on Windows. instead or a URL, the nub is addresses with
an argument vector. here is a sketch of the interface:

/*
 * const char* argv[] = { "Xi", "--nub-server" };
 *
 * xi_nub_server *server = xi_nub_server_new(argc, argv);
 * xi_nub_accept(server, myaccept_cb);
 *
 * xi_nub_client *client = xi_nub_client_new(argc, argv)
 * xi_nub_connect(client, myconnect_cb)
 *
 * xi_nub_conn_read(conn, buf, len, myread_cb)
 * xi_nub_conn_write(conn, buf, len, mywrite_cb)
 * xi_nub_conn_close(conn, myclose_cb)
 */