xi-nub: improve reliability for closed named pipes on windows
A reliability issues exists on windows in the situation where
the server writes to a pipe and subsequently calls disconnect.

- Write
- DisconnectNamedPipe

and a client calls read:

- ReadFile
- GetOverlappedResult

The client spuriously receives ERROR_PIPE_NOT_CONNECTED and it is
due to DisconectNamedPipe being processed before data from the
WriteFile call is read. ReadFile responds to DisconnectNamedPipe
and the data in the buffer is discarded. It is a race and differs
from graceful shutdown after write on POSIX sockets, whereby
operations are ordered, rather here the pipe is 'forced' closed.

This is problematic, because there is no other graceful 'shutdown'
method allowing one to disconnect a named pipe and then reuse it,
because CloseHandle completely closes the socket.

In resolution, this patch introduces a command byte with a 'quit'
opcode (0) and a 'request' opcode (1). The server issues another
ReadFile after each operation, and only calls DisconnectNamedPipe,
in response to the 'quit' opcode. The client then can always read
the response from the server without worry of being disconnected
before reading the response.

It seems an additional round-trip or delay is necessary for reliable
operation, because it is necessary to call DisconnectNamedPipe to
reuse the pipe for a successive connection. DisconnectNamedPipe is
somewhat defensively documented in Windows Developer documentation
as 'forcing' the end of the named pipe closed.