xi: enable nub search by default, and add --stand-alone flag
the nub search is now reliable enough to be enabled by default.
this has been achieved by routing the "unicode" command via a nub
client and the "nub-client" method has been moved to a command line
flag named `--stand-alone` which is used in conjunction with the
"unicode" sub-command e.g.

  xi --stand-alone --timings unicode smiling cat
  [Rabin-Karp] load = 178645μs, search = 34μs
  😸    U+1f638  GRINNING CAT FACE WITH SMILING EYES
  😺    U+1f63a  SMILING CAT FACE WITH OPEN MOUTH
  😻    U+1f63b  SMILING CAT FACE WITH HEART-SHAPED EYES

- the "unicode" subcommand now does a remote search.
- add the ability to run server with the `--brute-force` flag.
- make stand alone search available via the `--stand-alone` flag.
- the `--brute-force` flag can be combined with the `--stand-alone`
  flag to run the unoptimized linear scan search.