xi-nub: virtualize nub app name used for session object names
the focus of this change is to add an app_name parameter to
the xi_nub_ctx object and for nub methods to use this app name
instead of hard coded values.

some profile objects were hardcoded to "Xi\" or "Xi-". they have
been changed to "%s\\" or "%s-" plus the app_name from the context.

the following public interfaces have changed:

- remove xi_nub_init_ctx(xi_nub_ctx *ctx);
- add xi_nub_ctx* xi_nub_ctx_create(const char *app_name);
- add void xi_nub_ctx_destroy(xi_nub_ctx *ctx);

the apps have been changed in the following way:

- do_nub_client, do_nub_server and test_sem call xi_nub_ctx_create
  and xi_nub_destroy instead of xi_nub_get_root_context.
- the application profile directory templates have been changed
  to include the app name.
- semaphores, unix socket and named pipes now use the  app name
  from the context instead of hard coded strings.
- xi_nub_get_root_context now returns a pointer to the first
  object created with xi_nub_ctx_create.
- unix sockets functions are part of xi-common and do not
  have xi_nub_cxt parameters, so in one place for unix socket
  addresses on macos, xi_nub_get_root_context is used.